javaBean类型 Signed-off-by: caihe <caihe@gintong.com>

+++ phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/model/ColumnKnowledge.java
@@ -14,7 +14,7 @@ public class ColumnKnowledge implements Serializable {
 	private static final long serialVersionUID = -5440480908931909931L;
 	private long column_id;// 栏目ID
 	private long knowledge_id;// 知识ID
-	private int user_id;// 用户id
+	private long user_id;// 用户id
 	private int type;// 类型
 
 	public long getColumn_id() {
@@ -33,11 +33,11 @@ public class ColumnKnowledge implements Serializable {
 		this.knowledge_id = knowledge_id;
 	}
 
-	public int getUser_id() {
+	public long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(int user_id) {
+	public void setUser_id(long user_id) {
 		this.user_id = user_id;
 	}
